Надеюсь финальная фиксация ;)

+++ hw06-spring-jpql/src/test/java/krg/petr/otusru/utils/TestDataGenerator.java
@@ -6,6 +6,7 @@ import krg.petr.otusru.models.Genre;
 
 import java.util.List;
 import java.util.stream.IntStream;
+import java.util.stream.LongStream;
 
 public final class TestDataGenerator {
 
@@ -14,13 +15,13 @@ public final class TestDataGenerator {
     }
 
     public static List<Author> getDbAuthors() {
-        return IntStream.range(1, 4).boxed()
+        return LongStream.range(1, 4).boxed()
                 .map(id -> new Author(id, "Author_" + id))
                 .toList();
     }
 
     public static List<Genre> getDbGenres() {
-        return IntStream.range(1, 7).boxed()
+        return LongStream.range(1, 7).boxed()
                 .map(id -> new Genre(id, "Genre_" + id))
                 .toList();
     }
